requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2-binary==2.9.3
django-environ==0.9.0
gunicorn==20.1.0
dj-database-url==1.0.0